Added 'hours' & 'date' as optional fields in the exports
Added both hours and date to both export types.
Choose the 'string' datatype as that makes it easiest to fit in the way the project is structured.
Both flags are optional and by default disabled.

Build successfully with 'make VERSION=0.0.8' and tested both outputs with the optional flags. -> Output as desired.